Adding templates for README and CHANGELOG
Previously these were appendices in the online documentation, and so whilst changes to the templates were tracked in document history - this is not particularly accessible, or consistent with how changes are tracked for other elements of the project template.

Given the importance of the README specifically this was not ideal. Instead a '.template' convention is now followed to indicate where a file has a template version which should replace the similarly named instance of the file.

For example the README file has a template, which is generic, and a instance, which is the README describing the base project itself.