Adding a setup script for getting started quicker
This automated a lot of the boring stuff like resetting the git repository, creating an initial commit and switching to a develop branch, and performing all the various find and replace operations across the project.

This script is intended to compliment the manual process, which will still need to be followed for all but the most common use-cases.

There are likely to be bugs or edge cases this script doesn't cover. Whilst I am happy to merge fixes for these, I won't be particularly proactive in finding them, instead opting to patch things up afterwards.

In particular the script doesn't cope with being replayed (say you make a mistake and won't to change one of the find/replace answers). In these cases it's expected you will just nuke the directory and start again. Since this is only intended for new projects I don't consider that a significant limitation.