Adding migration guide to move between versions
It was becoming increasingly difficult to remember the difference between versions, in order to incrementally move between Pristine versions.

To address this, and better support existing projects, a migration guide will now document any changes that need to be made to Pristine (except where these clash with per-project customisations).

This migration guide is removed automatically as part of the setup process since it's a 'meta' document and not expected to be maintained for individual projects (though this is not precluded).